Changed file structure

+++ src/com/myshopnet/models/Cart.java
@@ -6,7 +6,9 @@ import java.util.List;
 
 public class Cart {
     private final List<CartItem> items = new ArrayList<>();
-    public void add(String sku, int qty) { items.add(new CartItem(sku, qty)); }
+    public void add(String sku, Long qty) {
+        items.add(new CartItem(sku, qty));
+    }
     public List<CartItem> items() { return Collections.unmodifiableList(items); }
 }
